Initial setup for postgis container
This commit introduces initial support for creating a
PostGIS enabled database container. By default the container
will accept connections from any IP using md5.

Users can use environment variables to initialize a PostGIS
database:
 - POSTGRES_USER: user to create
 - POSTGRES_PASSWORD: password to use for database
 - POSTGRES_DB: database to create and enable PostGIS on

This container is intended to be used for development purposes and
should not be used in production environments without modification
and tuning. In particular, this container eschews flexibility in
runtime configuration in favor of readability and maintainability since
it is intended to be used only in local development environments.