Fix package version references, add test, fix docs
Update the package version references for PostgreSQL and PostGIS. In the
process, bump the version of PostGIS to 2.2.2. Also, update the README
so that the examples work.

Lastly, add a test that actually exercises the PostGIS installation a
bit to ensure that the image has the bits that it needs.